Switch to normal sessions, refactor.
There's no point in JWTs and just sticking sessions in a
database is a lot easier. We're also storing the session
ID in a HTTP-only cookie, which means it'll be sent on
images so we can back out the encryption of photo keys, too.

As part of this big refactoring, ditch sqlx in favour of
upper/db as it handles nested structs and inserts way
nicer.