Breaking: refactor storage drivers
Move the interfaces to the callers; make the store accept a 'type'
of object (photo, thumbnail, etc) instead of using two stores.

This changes the path on disk that photos are stored at; anyone
brave enough to be running Simpic at this stage should clear
their data directory and execute `TRUNCATE photos CASCADE`
on the database.