Make lightbox loading awesome.
Introduces a promise-based LRU cache for thumbnails, metadata and
images, and also send caching headers from the backend.

When the lightbox appears it first gets the metadata (which
contains width/height as of the recent ingress changes), then
gets the thumbnail and renders it blurred and the same size as
the actual image will be, then gets the full image and renders
that.

In most cases the metadata and thumbnail are going to be cached
so they will be instant. When loading from scratch these will
be in the hot path but it's worth it to show *something* while
the potentially-giant image loads.

All in all this gives a progressive loading effect, and makes
the lightbox feel a lot snappier.

Closes #39